+++ shared/src/commonMain/kotlin/business/datasource/network/main/responses/BasketDeleteRequestDTO.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BasketDeleteRequestDTO(
-    @SerialName("product") val product: Int,
+    @SerialName("product") val product: Long,
 )
