+++ shared/src/commonMain/kotlin/business/datasource/network/main/responses/BuyRequestDTO.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BuyRequestDTO(
-    @SerialName("address_id")  val addressId: Int,
+    @SerialName("address_id")  val addressId: Long,
     @SerialName("shipping_type")  val shippingType: Int,
     )
